src/requirements.txt
Flask==2.3.2
Flask_Cors==3.0.10
flask_sqlalchemy==3.0.2
python-dotenv==0.21.1
python_jose==3.3.0
Werkzeug==2.3.3
psycopg2==2.9.9
gunicorn==21.1.0
